Consolidate obsolete forwarding shims into src/Obsolete.V4.cs
Move the Reusable, Pruning, Seeking, Sorting, and StringOut obsolete
delegate classes, plus the shadowed IndicatorConfigExtensions.ToBuilder
extension, out of their live *Extensions files and into a single
Obsolete.V4.cs at the src root, mirroring the existing
Obsolete.V3.Indicators.cs / Obsolete.V3.Other.cs pattern. This keeps
the deprecation compatibility layer segregated from ongoing
development, matching repo convention (including the RCS1141/RCS1142
doc-comment suppression already used by Obsolete.V3.Other.cs).
