+++ src/Common/Catalog/Schema/IndicatorConfigExtensions.cs
@@ -1,32 +1,10 @@
-using System.Diagnostics.CodeAnalysis;
-
 namespace FacioQuo.Stock.Indicators;
 
 /// <summary>
 /// Extension methods for working with indicator configurations.
 /// </summary>
-public static class IndicatorConfigExtensions
+public static partial class IndicatorConfigExtensions
 {
-    /// <summary>
-    /// Obsolete. Permanently shadowed by the <see cref="IndicatorConfig.ToBuilder"/> instance
-    /// method, which always takes precedence over this same-named extension method. Removal
-    /// tracked in issue #2139.
-    /// </summary>
-    /// <param name="config">Indicator configuration.</param>
-    /// <returns>A <see cref="ListingExecutionBuilder"/>.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when config is null.</exception>
-    [ExcludeFromCodeCoverage]
-    [Obsolete("This extension method is permanently shadowed by the `IndicatorConfig.ToBuilder()` instance method and will be removed.", false)]
-    public static ListingExecutionBuilder ToBuilder(this IndicatorConfig config)
-    {
-        ArgumentNullException.ThrowIfNull(config);
-
-        // not recursive: C# overload resolution binds this call to the
-        // IndicatorConfig.ToBuilder() instance method, which always takes
-        // precedence over this same-named extension method
-        return config.ToBuilder();
-    }
-
     /// <summary>
     /// Executes an indicator configuration with the provided bars.
     /// </summary>
